==> research/01_data_ingestion.ipynb <==
import os
%pwd

os.chdir("../")
%pwd

# Entity
from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class DataIngestionConfig:
    root_dir: Path
    raw_data_path: Path
    train_data_path: Path
    test_data_path: Path    

from customer_churn.constants import *
from customer_churn.utils.common_utils import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)

        create_directories([self.config.artifacts_root])

    def get_data_ingestion_config(self) -> DataIngestionConfig:
        config = self.config.data_ingestion

        create_directories([config.root_dir])

        data_ingestion_config = DataIngestionConfig(
            root_dir=config.root_dir,
            raw_data_path=config.raw_data_path,
            train_data_path=config.train_data_path,
            test_data_path=config.test_data_path 
        )

        return data_ingestion_config

import os
import sys
import pandas as pd
from sklearn.model_selection import train_test_split
from dataclasses import dataclass
from box import ConfigBox
from box.exceptions import BoxValueError
from customer_churn import logger
from customer_churn.utils.common_utils import read_sql_data

class DataIngestion:
    def __init__(self, config: DataIngestionConfig):
        self.config = config

    def initiate_data_ingestion(self):
        logger.info("Entered the Data Ingestion")
        try:
            #df = pd.read_csv("notebook/data/EDA.csv")
            df = read_sql_data()
            logger.info("Read the data as DataFrame from MySQL.")


            os.makedirs(os.path.dirname(self.config.raw_data_path), exist_ok=True)

            
            df.to_csv(self.config.raw_data_path, index=False, header=True)
            logger.info("Reading Completed.")

           
            logger.info("train_test_split initiated...............")
            train_set, test_set = train_test_split(df, test_size=0.2, random_state=42)

            train_set.to_csv(self.config.train_data_path, index=False, header=True)
            test_set.to_csv(self.config.test_data_path, index=False, header=True)
            logger.info("data ingestion completed.")

            return(
                self.config.train_data_path,
                self.config.test_data_path
            )

        except BoxValueError:
            raise ValueError("Error occured in Data Ingestion...............")
        except Exception as e:
            raise e

try:
    config = ConfigurationManager()
    data_ingestion_config = config.get_data_ingestion_config()
    data_ingestion = DataIngestion(config=data_ingestion_config)
    data_ingestion.initiate_data_ingestion()
except Exception as e:
    raise e
